Final fix of the game start
