+++ Assets/Python/Handlers.py
@@ -4,7 +4,6 @@ import ScreenHandler
 import CityNameManager
 import Civilizations
 import History
-import Locations
 import Messages
 import Modifiers
 import Rules
